refit in sep prepare node
